macOS 10.15 beta 4, Xcode 11 beta 4 (iOS, tvOS, watchOS, iOSSupport 13)

+++ iOS/PrivateFrameworks/PhotosUICore.framework/PXForYouSuggestionDisplayAsset.h
@@ -64,7 +64,7 @@
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, copy) NSString *description;
 @property(readonly) unsigned long long hash;
-@property(readonly, nonatomic) float hdrGain;
+@property(readonly, nonatomic) double hdrGain;
 @property(readonly, nonatomic) NSDate *importDate;
 @property(readonly, nonatomic) NSString *localizedGeoDescription;
 @property(readonly) Class superclass;
