+++ iOS/PrivateFrameworks/iWorkImport.framework/TSCHChartFeature.h
@@ -84,11 +84,11 @@ __attribute__((visibility("hidden")))
 - (id)valueLabelPositioner;
 - (id)defaultDataFileName;
 - (unsigned long long)styleIndexForAxisID:(id)arg1;
-- (Class)valueAxisClassForID:(id)arg1 scale:(int)arg2;
+- (Class)valueAxisClassForID:(id)arg1 scale:(long long)arg2;
 - (void)p_initializeAxisIDs;
 - (id)filteredStyleOwnersFromStyleOwners:(id)arg1;
 - (_Bool)supportsEditingForAxisID:(id)arg1;
-- (Class)p_selectAxisClassForScale:(int)arg1 linearClass:(Class)arg2 logClass:(Class)arg3 percentClass:(Class)arg4;
+- (Class)p_selectAxisClassForScale:(long long)arg1 linearClass:(Class)arg2 logClass:(Class)arg3 percentClass:(Class)arg4;
 - (void)dealloc;
 - (void)clearParent;
 - (id)initWithChartType:(id)arg1;
