+++ iOS/PrivateFrameworks/iWorkImport.framework/TSCHChartFeatureBar.h
@@ -32,7 +32,7 @@ __attribute__((visibility("hidden")))
 - (id)genericToSpecificPropertyMap;
 - (id)defaultDataFileName;
 - (Class)presetImagerClass;
-- (Class)valueAxisClassForID:(id)arg1 scale:(int)arg2;
+- (Class)valueAxisClassForID:(id)arg1 scale:(long long)arg2;
 - (void)p_initializeAxisIDs;
 - (int)deprecated3DShadowSpecificProperty;
 - (int)deprecated3DBevelEdgesSpecificProperty;
