+++ iOS/PrivateFrameworks/iWorkImport.framework/TSCHChartFeatureMixed.h
@@ -23,7 +23,7 @@ __attribute__((visibility("hidden")))
 - (id)genericToSpecificPropertyMap;
 - (id)defaultDataFileName;
 - (Class)presetImagerClass;
-- (Class)valueAxisClassForID:(id)arg1 scale:(int)arg2;
+- (Class)valueAxisClassForID:(id)arg1 scale:(long long)arg2;
 
 @end
