+++ iOS/PrivateFrameworks/iWorkImport.framework/TSDColorFill.h
@@ -56,7 +56,7 @@ __attribute__((visibility("hidden")))
 - (long long)mixingTypeWithObject:(id)arg1 context:(id)arg2;
 - (id)referenceColor;
 - (id)presetKind;
-- (int)fillType;
+- (long long)fillType;
 - (void)paintPath:(struct CGPath *)arg1 inContext:(struct CGContext *)arg2;
 - (void)paintRect:(struct CGRect)arg1 inContext:(struct CGContext *)arg2;
 - (_Bool)drawsInOneStep;
