+++ iOS/PrivateFrameworks/iWorkImport.framework/TSDShapeRenderingOperation.h
@@ -11,7 +11,7 @@
 __attribute__((visibility("hidden")))
 @interface TSDShapeRenderingOperation : TSDSwatchRenderingOperation
 {
-    int mShapeType;
+    long long mShapeType;
     TSDPathSource *mShapePathSource;
     double mAngle;
     unsigned long long mBackgroundAppearance;
@@ -23,8 +23,8 @@ __attribute__((visibility("hidden")))
 - (void)doWorkWithReadLock;
 - (_Bool)needsPressedStateBackground;
 - (struct UIEdgeInsets)swatchEdgeInsets;
-- (id)initWithPreset:(id)arg1 imageSize:(struct CGSize)arg2 imageScale:(double)arg3 swatchFrame:(struct CGRect)arg4 shapeType:(int)arg5 shapePathSource:(id)arg6 angle:(double)arg7 backgroundAppearance:(unsigned long long)arg8 documentRoot:(id)arg9;
-- (id)initWithPreset:(id)arg1 imageSize:(struct CGSize)arg2 imageScale:(double)arg3 swatchFrame:(struct CGRect)arg4 shapeType:(int)arg5 shapePathSource:(id)arg6 documentRoot:(id)arg7;
+- (id)initWithPreset:(id)arg1 imageSize:(struct CGSize)arg2 imageScale:(double)arg3 swatchFrame:(struct CGRect)arg4 shapeType:(long long)arg5 shapePathSource:(id)arg6 angle:(double)arg7 backgroundAppearance:(unsigned long long)arg8 documentRoot:(id)arg9;
+- (id)initWithPreset:(id)arg1 imageSize:(struct CGSize)arg2 imageScale:(double)arg3 swatchFrame:(struct CGRect)arg4 shapeType:(long long)arg5 shapePathSource:(id)arg6 documentRoot:(id)arg7;
 
 @end
