+++ iOS/PrivateFrameworks/iWorkImport.framework/TSWPStorageBroadcaster.h
@@ -19,8 +19,8 @@ __attribute__((visibility("hidden")))
 - (void)broadcastStorage:(id)arg1 didChangeParagraphsInIndexRange:(struct _NSRange)arg2;
 - (void)broadcastStorage:(id)arg1 didDeleteParagraphsInIndexRange:(struct _NSRange)arg2;
 - (void)broadcastStorage:(id)arg1 didInsertParagraphsInIndexRange:(struct _NSRange)arg2;
-- (void)broadcastStorage:(id)arg1 didChangeRange:(struct _NSRange)arg2 delta:(long long)arg3 broadcastKind:(int)arg4 attributeKindChanges:(id)arg5;
-- (void)broadcastStorage:(id)arg1 didChangeRange:(struct _NSRange)arg2 delta:(long long)arg3 broadcastKind:(int)arg4;
+- (void)broadcastStorage:(id)arg1 didChangeRange:(struct _NSRange)arg2 delta:(long long)arg3 broadcastKind:(unsigned long long)arg4 attributeKindChanges:(id)arg5;
+- (void)broadcastStorage:(id)arg1 didChangeRange:(struct _NSRange)arg2 delta:(long long)arg3 broadcastKind:(unsigned long long)arg4;
 - (void)removeAllObservers;
 - (void)removeParagraphObserver:(id)arg1;
 - (void)addParagraphObserver:(id)arg1;
