+++ iOS/PrivateFrameworks/iWorkImport.framework/TSWPStorageObserver-Protocol.h
@@ -11,6 +11,6 @@
 @protocol TSWPStorageObserver <NSObject>
 
 @optional
-- (void)storage:(TSWPStorage *)arg1 didChangeRange:(struct _NSRange)arg2 delta:(long long)arg3 broadcastKind:(int)arg4 attributeKindChanges:(NSArray *)arg5;
+- (void)storage:(TSWPStorage *)arg1 didChangeRange:(struct _NSRange)arg2 delta:(long long)arg3 broadcastKind:(unsigned long long)arg4 attributeKindChanges:(NSArray *)arg5;
 @end
